+++ adapter/src/main/java/me/googas/chat/debug/DebugTask.java
@@ -1,7 +1,3 @@
 package me.googas.chat.debug;
 
-
-public interface DebugTask {
-
-
-}
+public interface DebugTask {}